move test_get_compromised_nodes into its own file
